commit changes

+++ src/main/java/net/dirtcraft/discord/discordlink/Commands/Discord/StopServer.java
@@ -1,5 +1,6 @@
 package net.dirtcraft.discord.discordlink.Commands.Discord;
 
+import net.dirtcraft.discord.discordlink.API.Channels;
 import net.dirtcraft.discord.discordlink.API.MessageSource;
 import net.dirtcraft.discord.discordlink.Commands.DiscordCommandExecutor;
 
@@ -10,8 +11,9 @@ public class StopServer implements DiscordCommandExecutor {
     @Override
     public void execute(MessageSource source, String command, List<String> args) {
         try {
-            source.getGamechat().sendEmbed("Discord-Link Reboot", "Attempting to reboot the server.");
-            source.getMessage().delete().queue(s->{},e->{});
+            if (!source.isPrivateMessage()) source.getMessage().delete().queue(s->{},e->{});
+            else source.sendCommandResponse("Forced Reboot Scheduled.", "Attempting to reboot the server.");
+            Channels.getDefaultChat().sendMessage(source, "Forced Reboot Scheduled.", "Attempting to reboot the server.");
             Thread.sleep(555);
         } catch (Throwable ignored){
